refactor: move middlewares to own file
See: BEDS-605
